Add ability to pass a custom user agent fragment (#56)
The fragment will be appended at the very end of the default user-agent. This enables consumers of the Terraform provider to pass custom tokens to perform additional analytics.

As an example, we will be adding a custom token in the [Consul/DNSimple integration](https://github.com/dnsimple/terraform-dnsimple-cts) to be able to better monitor the usage of the Consul Integration.