Document how to migrate a deprecated resource in the form of a guide (#69)
This PR introduces a migration guide for deprecated resources. The guide is intended to be generic with a detailed example. The example used in the guide is the migration of dnsimple_record to dnsimple_zone_record which is an actual deprecation change in the provider.

The intended use of the guide going forward is to include a link to it when we are deprecating a resource that has a replacement.