Implement Zone resource (#184)
We are adding a zone resource in favor of the `dnsimple_zone` data source, so we can add support for the zone active state management. With the new resource, developers can deactivate or activate zones, essentially managing whether a zone is publicly resolving or not. More on the API endpoints used [here](https://developer.dnsimple.com/v2/zones/#activateZoneService).