Trim trailing dot (root) from name servers in the `dnsimple_domain_delegation` resource (#203)
It was brought up some time ago that having fully qualified name server entries (i.e. with a . at the end) would result in perma diff since our API normalizes the entries and returns the normalized entries creating the difference. Since this is a known behaviour it made sense for the provider to acknowledge it and handle this case.